refactor(settings): WorkTab + _shared-Helfer extrahiert (H4 Schritt 2)
